[FEAT] UseCaseProtocol 추가 #30
