+++ POLZZAK/Coupon/CouponList/CouponListViewModel.swift
@@ -9,7 +9,7 @@ import Foundation
 import Combine
 
 final class CouponListViewModel: TabFilterViewModelProtocol, PullToRefreshProtocol, LoadingViewModelProtocol {
-    private let useCase: CouponsUsecase
+    private let useCase: CouponsUseCase
     
     var dataList = CurrentValueSubject<[CouponList], Never>([])
     var cancellables = Set<AnyCancellable>()
@@ -28,7 +28,7 @@ final class CouponListViewModel: TabFilterViewModelProtocol, PullToRefreshProtoc
     var dataDeleted = CurrentValueSubject<IndexPath?, Never>(nil)
     var showErrorAlertSubject = PassthroughSubject<Error, Never>()
     
-    init(useCase: CouponsUsecase) {
+    init(useCase: CouponsUseCase) {
         self.useCase = useCase
         
         //TODO: - DTO에서 Model로 변환할때 UserType을 단순하게 부모인지 아이인지 변환하고 UserInfo에서 사용하는 Model에 userType을 추가했으면 좋겠음.
