+++ POLZZAK/DomainLayer/UseCases/Protocol/LinkRequestUseCase.swift
